requirements.txt
numpy>=1.24.0
numba>=0.58.0
psycopg2-binary>=2.9.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
